pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "fastapi>=0.128.0",
    "isort>=7.0.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
